+++ src/egeo-demo/st-form-demo/json-schema.ts
@@ -16,94 +16,6 @@ export const JSON_SCHEMA: any = {
       "component": "standard"
    },
    "properties": {
-      "a": {
-         "title": "a",
-         "required": true,
-         "readOnly": false,
-         "type": "string",
-         "default": "",
-         "level": 1,
-         "exclusiveMinimum": false,
-         "exclusiveMaximum": false,
-         "internalName": "SERVICE_ID",
-         'ui': { 'relatedTo': 'b'}
-      },
-      "b": {
-         "title": "b",
-         "required": true,
-         "readOnly": false,
-         "type": "string",
-         "enum": ['a', '/sparta/sparta-server/null', 'c'],
-         "default": "/sparta/sparta-server/null",
-         "level": 1,
-         "exclusiveMinimum": false,
-         "exclusiveMaximum": false,
-         "internalName": "SERVICE_CONFIGURATION"
-      },
-      "c": {
-         "title": "c",
-         "required": true,
-         "readOnly": false,
-         "type": "string",
-         "default": "",
-         "level": 1,
-         "exclusiveMinimum": false,
-         "exclusiveMaximum": false,
-         "internalName": "SERVICE_ID",
-         'ui': {
-            'relatedTo': 'd'
-         }
-      },
-      "d": {
-         "title": "d",
-         "required": true,
-         "readOnly": false,
-         "type": "string",
-         "enum": ['a', '/sparta/sparta-server/null', 'c'],
-         "default": "/sparta/sparta-server/null",
-         "level": 1,
-         "exclusiveMinimum": false,
-         "exclusiveMaximum": false,
-         "internalName": "SERVICE_CONFIGURATION",
-         'ui': { 'relatedTo': 'e' }
-      },
-      "e": {
-         "title": "e",
-         "required": true,
-         "readOnly": false,
-         "type": "string",
-         "enum": ['a', '/sparta/sparta-server/null', 'c'],
-         "default": "/sparta/sparta-server/null",
-         "level": 1,
-         "exclusiveMinimum": false,
-         "exclusiveMaximum": false,
-         "internalName": "SERVICE_CONFIGURATION"
-      },
-      "f": {
-         "title": "f",
-         "required": true,
-         "readOnly": false,
-         "type": "string",
-         "enum": ['a', '/sparta/sparta-server/null', 'c'],
-         "default": "/sparta/sparta-server/null",
-         "level": 1,
-         "exclusiveMinimum": false,
-         "exclusiveMaximum": false,
-         "internalName": "SERVICE_CONFIGURATION",
-         'ui': { 'relatedTo': 'g' }
-      },
-      "g": {
-         "title": "g",
-         "required": true,
-         "readOnly": false,
-         "type": "string",
-         "enum": ['a', '/sparta/sparta-server/null', 'c'],
-         "default": "/sparta/sparta-server/null",
-         "level": 1,
-         "exclusiveMinimum": false,
-         "exclusiveMaximum": false,
-         "internalName": "SERVICE_CONFIGURATION"
-      },
       "general": {
          "type": "object",
          "title": "General",
@@ -112,29 +24,51 @@ export const JSON_SCHEMA: any = {
             "component": "show-more"
          },
          "properties": {
-            "serviceId": {
-               "title": "Name",
-               "required": true,
-               "readOnly": false,
-               "type": "string",
-               "default": "",
-               "level": 1,
-               "exclusiveMinimum": false,
-               "exclusiveMaximum": false,
-               "internalName": "SERVICE_ID"
-            },
             "serviceConfiguration": {
                "title": "Type of Sparta",
-               "required": true,
-               "readOnly": false,
+               "type": "string",
+               "enum": ['a', '/sparta/sparta-server/null', 'c'],
+               "examples": ['/sparta/sparta-server/null'],
+               "default": "/sparta/sparta-server/null",
+               'ui': { 'relatedTo': 'serviceId' }
+            },
+            "serviceId": {
+               "title": "Name",
+               "type": "string",
+               "default": ""
+            },
+
+            "a": {
+               "title": "a",
+               "type": "string",
+               'ui': { 'relatedTo': 'b'}
+            },
+            "b": {
+               "title": "b",
+               "type": "string",
+               "enum": ['a', '/sparta/sparta-server/null', 'c'],
+               "default": "/sparta/sparta-server/null"
+            },
+            "c": {
+               "title": "c",
+               "type": "string",
+               "default": "",
+               'ui': {
+                  'relatedTo': 'd'
+               }
+            },
+            "d": {
+               "title": "d",
                "type": "string",
                "enum": ['a', '/sparta/sparta-server/null', 'c'],
                "default": "/sparta/sparta-server/null",
-               "level": 1,
-               "exclusiveMinimum": false,
-               "exclusiveMaximum": false,
-               "internalName": "SERVICE_CONFIGURATION",
-               'ui': { 'relatedTo': 'serviceId' }
+               'ui': { 'relatedTo': 'e' }
+            },
+            "e": {
+               "title": "e",
+               "type": "string",
+               "enum": ['a', '/sparta/sparta-server/null', 'c'],
+               "default": "/sparta/sparta-server/null"
             },
             "runMode": {
                "title": "Run mode",