+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## 9.0.0 (upcoming)
 
+**New features:**
+
 * st-form: Allow to display and hide fields according to the value of another fields
+
+**Others**
+
+* Apply visual changes to several components for contrast improvements
 
 
 ## 8.8.0 (upcoming)