+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 8.8.0 (upcoming)
+## 8.8.0 (April 04, 2018)
 
 **New features:**